# openmeter-collector/builder-api/cmd/builder-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clearinghouse Builder API",
    "version": "1.0.0",
    "description": "Machine-facing Builder API for clearinghouse integrators. Provisions Auth0 end-users, OpenMeter customers, and mints short-lived signer JWTs."
  },
  "servers": [
    {
      "url": "http://localhost:8095",
      "description": "Local builder-api (openmeter-collector container)"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/apps/{clientId}/users": {
      "post": {
        "summary": "Create or upsert an end-user",
        "description": "Creates an Auth0 Database user, upserts the OpenMeter customer, and optionally issues an end-user API key.",
        "security": [
          {
            "tenantBasic": []
          },
          {
            "m2mBasic": []
          }
        ],
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Public Auth0 client id for the integrator app"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserResponse"
                }
              }
            }
          },
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/apps/{clientId}/oidc/token": {
      "post": {
        "summary": "RFC 8693 signer session token exchange",
        "description": "Exchanges an Auth0 end-user access token (device code / authorization code) or end-user API key (`sk_*`) for a short-lived signer JWT. The `{clientId}` path segment is the public Auth0 app client id. Authenticate with the end-user `subject_token`; optional HTTP Basic with the signer M2M client is supported for server-side callers.",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Public Auth0 client id for the integrator app"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/TokenExchangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signer session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignerSession"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, grant, target, or unsupported token type"
          },
          "401": {
            "description": "Invalid client credentials"
          }
        }
      }
    },
    "/api/v1/oidc/token": {
      "post": {
        "summary": "RFC 8693 signer session token exchange (client inferred)",
        "description": "Same exchange as `/api/v1/apps/{clientId}/oidc/token`, but without a path client id. The target app is inferred from the `subject_token`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/TokenExchangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signer session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignerSession"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, grant, target, or unsupported token type"
          },
          "401": {
            "description": "Invalid client credentials"
          }
        }
      }
    },
    "/api/v1/apps/{clientId}/usage": {
      "get": {
        "summary": "Query metered usage for an app",
        "description": "Subjects are derived from the authorized clientId, never from caller input. Without externalUserId every customer of the app is included.",
        "security": [
          {
            "tenantBasic": []
          },
          {
            "m2mBasic": []
          }
        ],
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "meter",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "externalUserId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "groupBy",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usage rows scoped to the app"
          },
          "400": {
            "description": "Invalid meter, time window, or externalUserId"
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "404": {
            "description": "Unknown app, or the caller does not own it"
          },
          "502": {
            "description": "Metering backend unavailable"
          },
          "503": {
            "description": "Metering backend not configured"
          }
        }
      }
    },
    "/api/v1/users/me/usage": {
      "get": {
        "summary": "Query metered usage for the authenticated user",
        "description": "User-scoped usage route. Identity is derived from the Bearer signer JWT; the app client id is inferred from token claims.",
        "security": [
          {
            "signerBearer": []
          }
        ],
        "parameters": [
          {
            "name": "meter",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "groupBy",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usage rows scoped to the authenticated user"
          },
          "400": {
            "description": "Invalid meter or time window"
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "502": {
            "description": "Metering backend unavailable"
          },
          "503": {
            "description": "JWT verifier or metering backend is not configured"
          }
        }
      }
    },
    "/api/v1/apps/{clientId}/users/{externalUserId}/access": {
      "get": {
        "summary": "Read a user's allowance / entitlement balance",
        "security": [
          {
            "tenantBasic": []
          },
          {
            "m2mBasic": []
          }
        ],
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "externalUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Must not contain ':' \u2014 customer keys are 'clientId:externalUserId'."
          },
          {
            "name": "feature",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Access snapshot"
          },
          "400": {
            "description": "Invalid externalUserId"
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "404": {
            "description": "Unknown app, or the caller does not own it"
          },
          "502": {
            "description": "Metering backend unavailable"
          },
          "503": {
            "description": "Metering backend not configured"
          }
        }
      }
    },
    "/api/v1/apps/{clientId}/users/{externalUserId}/grants": {
      "post": {
        "summary": "Grant a user allowance",
        "security": [
          {
            "tenantBasic": []
          },
          {
            "m2mBasic": []
          }
        ],
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "externalUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Must not contain ':' \u2014 customer keys are 'clientId:externalUserId'."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amountUsdMicros"
                ],
                "properties": {
                  "amountUsdMicros": {
                    "type": "integer",
                    "format": "int64",
                    "minimum": 1
                  },
                  "featureKey": {
                    "type": "string"
                  },
                  "grantKey": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Grant applied"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "404": {
            "description": "Unknown app, or the caller does not own it"
          },
          "502": {
            "description": "Grant failed"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "m2mBasic": {
        "type": "http",
        "scheme": "basic",
        "description": "M2M client id and secret (users:write, RFC 8693 token exchange)"
      },
      "tenantBasic": {
        "type": "http",
        "scheme": "basic",
        "description": "Tenant admin credentials: username is the tenant clientId, password is its secret from TENANT_ADMIN_KEYS. A tenant may only address its own clientId; the platform M2M credential (m2mBasic) may address any. A request for another tenant answers 404, not 403, so that a shared OpenMeter tenant does not confirm which client ids exist."
      },
      "signerBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Signer JWT minted by the RFC 8693 token-exchange endpoint."
      }
    },
    "schemas": {
      "CreateUserRequest": {
        "type": "object",
        "required": [
          "externalUserId"
        ],
        "properties": {
          "externalUserId": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "connection": {
            "type": "string"
          },
          "issueApiKey": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "CreateUserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "clientId": {
            "type": "string"
          },
          "externalUserId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "apiKey": {
            "type": "string",
            "description": "Returned once when issueApiKey is true"
          }
        }
      },
      "TokenExchangeRequest": {
        "type": "object",
        "required": [
          "grant_type",
          "subject_token",
          "subject_token_type"
        ],
        "properties": {
          "grant_type": {
            "type": "string",
            "enum": [
              "urn:ietf:params:oauth:grant-type:token-exchange"
            ]
          },
          "subject_token": {
            "type": "string",
            "description": "Auth0 user access token (JWT) or end-user API key (sk_*)"
          },
          "subject_token_type": {
            "type": "string",
            "enum": [
              "urn:ietf:params:oauth:token-type:access_token"
            ]
          },
          "requested_token_type": {
            "type": "string",
            "enum": [
              "urn:ietf:params:oauth:token-type:access_token"
            ]
          },
          "audience": {
            "type": "string",
            "description": "Must match configured signer audience when provided"
          },
          "resource": {
            "type": "string",
            "description": "Must match configured signer audience when provided"
          }
        }
      },
      "SignerSession": {
        "type": "object",
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "scope",
          "has_access",
          "balance_usd_micros"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "expires_in": {
            "type": "integer"
          },
          "scope": {
            "type": "string"
          },
          "signer_url": {
            "type": "string",
            "format": "uri"
          },
          "discovery_url": {
            "type": "string",
            "format": "uri"
          },
          "issued_token_type": {
            "type": "string"
          },
          "correlation_id": {
            "type": "string"
          },
          "has_access": {
            "type": "boolean",
            "description": "Whether the customer currently has allowance. Always present."
          },
          "balance_usd_micros": {
            "type": "integer",
            "format": "int64",
            "description": "Allowance balance in USD micros. Always present."
          }
        }
      }
    }
  }
}